Time: 0 ms (100.00%), Space: 5.9 MB (76.02%) - LeetHub

+++ 7-reverse-integer/7-reverse-integer.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     int reverse(int x) {
-        long long int last;
-        long long int temp;
+        int last;
+        int temp;
         long long int reverse;
         reverse=0;
         temp=x;
